requirements.txt
streamlit
pandas
pillow
numpy
fpdf
opencv-python-headless
easyocr
